move streamTransactioned to its own base class
