+++ backend/src/main/kotlin/org/loculus/backend/controller/InternalController.kt
@@ -18,9 +18,11 @@ import org.loculus.backend.api.ProcessedData
 import org.loculus.backend.api.SubmittedProcessedData
 import org.loculus.backend.api.UnprocessedData
 import org.loculus.backend.controller.LoculusCustomHeaders.X_TOTAL_RECORDS
+import org.loculus.backend.log.RequestIdContext
 import org.loculus.backend.model.RELEASED_DATA_RELATED_TABLES
 import org.loculus.backend.model.ReleasedDataModel
 import org.loculus.backend.service.submission.SubmissionDatabaseService
+import org.loculus.backend.utils.IteratorStreamer
 import org.springframework.http.HttpHeaders
 import org.springframework.http.HttpStatus
 import org.springframework.http.MediaType
@@ -46,7 +48,9 @@ private val log = KotlinLogging.logger { }
 open class InternalController(
     private val releasedDataModel: ReleasedDataModel,
     private val submissionDatabaseService: SubmissionDatabaseService,
-) {
+    iteratorStreamer: IteratorStreamer,
+    requestIdContext: RequestIdContext,
+) : BaseController(iteratorStreamer, requestIdContext) {
     @Operation(description = EXTRACT_UNPROCESSED_DATA_DESCRIPTION)
     @ApiResponse(
         responseCode = "200",
@@ -97,7 +101,7 @@ open class InternalController(
         val headers = HttpHeaders()
         headers.contentType = MediaType.parseMediaType(MediaType.APPLICATION_NDJSON_VALUE)
         headers.eTag = lastDatabaseWriteETag
-        val streamBody = submissionDatabaseService.streamTransactioned(
+        val streamBody = streamTransactioned(
             endpoint = "extract-unprocessed-data",
             organism = organism,
         ) {
@@ -232,7 +236,7 @@ open class InternalController(
         // We just need to make sure the etag used is from before the count
         // Alternatively, we could read once to file while counting and then stream the file
 
-        val streamBody = submissionDatabaseService.streamTransactioned(
+        val streamBody = streamTransactioned(
             compression,
             endpoint = "get-released-data",
             organism = organism,
